python-project-structure/requirements.txt
flask
dotenv
pytest
psycopg2-binary